services/websocket-service/pyproject.toml
[tool.poetry]
name = "websocket-service"
version = "0.1.0"
description = "WebSocket Service for Advanced Todo Chatbot"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
websockets = "^12.0"
kafka-python = "^2.0.2"
dapr-ext-fastapi = "^1.12.0"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"